pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "etf-agent-rotation"
version = "0.1.0"
description = "ETF动量轮动 + AI Agent辅助系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0,<2.3",
    "numpy>=1.24,<2.0",
    "akshare>=1.12.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
